% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Valiation_and_visualization.R
\name{ComENTR}
\alias{ComENTR}
\title{Entropy rate}
\usage{
ComENTR(
  entr_expr,
  entr_adja,
  entr_labels,
  local = TRUE,
  method = 1,
  quants.v = c(0.1, 0.9)
)
}
\arguments{
\item{entr_expr}{a genome-wide expression vector of a sample with names labeling the genes (also annotated to entrez gene IDs) of same length as rows of adj.m.}

\item{entr_adja}{an adjacency matrix representing a connected PPI network, with rownames/colnames of the matrix annotated to Entrez gene IDs.}

\item{entr_labels}{Classification (label) of the sample}

\item{local}{logical, if TRUE also compute the local Shannon (normalised) entropies.}

\item{method}{scalar specifying method to compute stochastic matrix: 1=stochastic vector of a node is independent of the node's gene expression value, 2=stochastic vector of a node is dependent on the node's gene expression value through use of the betaFn. Default is 1.}

\item{quants.v}{optionally, a vector of length 2 specifying the quantiles for defining low and high expression (only needed if method=2 is used).}
}
\value{
A boxplot about the entropy.
}
\description{
This function computes the entropy rate (SR) and visualization.
}
\examples{
data("entr_adja")
data("entr_expr")
data("entr_labels")
ComENTR(entr_expr,entr_adja,entr_labels,local=TRUE,method=1)

}
\references{
CompSR.R DoIntegPIN.R  Author: Andrew Teschendorff Date: 31 Dec 2013.
}
