% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_preprocessing.R
\name{data.filter}
\alias{data.filter}
\title{Data filter}
\usage{
data.filter(Data, percentage = 0.8)
}
\arguments{
\item{Data}{A matrix representing the genomic data such as gene expression data, miRNA expression data. \cr
For the matrix, the rows represent the genomic features, and the columns represent the samples.}

\item{percentage}{The maximum proportion in the sample with missing values}
}
\value{
The data matrix after filter.
}
\description{
Data filter for features with missing values
}
\examples{
Data=matrix(runif(1000),nrow = 50,ncol = 20)
geneName=paste("Gene", 1:50, sep = " ")
sampleName=paste("Sample", 1:20, sep = " ")
rownames(Data)=geneName
colnames(Data)=sampleName
index=sample(c(1:1000),300)
Data[index]=NA
result=data.filter(Data,percentage=0.6)
}
