% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_preprocessing.R
\name{data.normalization}
\alias{data.normalization}
\title{Data normalization}
\usage{
data.normalization(Data, type = "feature_Median", log2 = FALSE)
}
\arguments{
\item{Data}{A matrix representing the genomic data such as gene expression data, miRNA expression data. \cr
For the matrix, the rows represent the genomic features, and the columns represent the samples.}

\item{type}{A character value representing the normalization type. The optional values are shown below:
\itemize{
\item "feature_Median". The default value. Normalize dataset by sweeping the median values of each feature.
\item "feature_Mean". Normalize dataset by sweeping the mean values of each feature.
\item "feature_zscore". Conduct z_score normalization for each feature.
\item "sample_zscore".  Conduct z_score normalization for each samples.

}}

\item{log2}{A logical value. If TRUE, the data is transform as log2(x+1). This is commonly used for RNAseq data.}
}
\value{
The normalized data matrix.
}
\description{
Conduct normalization for dataset.
}
\examples{
data(mRNAexp)
result=data.normalization(mRNAexp,type="feature_Median",log2=FALSE)

}
\references{
Xu T, Le TD, Liu L, Su N, Wang R, Sun B, Colaprico A, Bontempi G, Li J. CancerSubtypes: an R/Bioconductor package for molecular cancer subtype identification, validation and visualization. Bioinformatics. 2017 Oct 1;33(19):3131-3133. doi: 10.1093/bioinformatics/btx378. PMID: 28605519.
}
