% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{entr_adja}
\alias{entr_adja}
\title{Dataset: entr_adja}
\format{
A matrix
}
\description{
an adjacency matrix representing a connected PPI network, with rownames/colnames of the matrix annotated to Entrez gene IDs.
}
\examples{
data(entr_adja)
}
\keyword{datasets}
