% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{entr_expr}
\alias{entr_expr}
\title{Dataset: entr_expr}
\format{
A matrix
}
\description{
a genome-wide expression vector of a sample with names labeling the genes (also annotated to entrez gene IDs) of same length as rows of adj.m.
}
\examples{
data(entr_expr)
}
\keyword{datasets}
