% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cluster_method.R
\name{ExecuteCC}
\alias{ExecuteCC}
\title{Execute Consensus Clustering based on five datasets.}
\usage{
ExecuteCC(
  clusterNum,
  mRNAexp,
  miRNAexp,
  lncRNAexp,
  methylation,
  CNVmatrix,
  maxK = 10,
  clusterAlg = "hc",
  distance = "pearson",
  title = "ConsensusClusterResult",
  reps = 500,
  pItem = 0.8,
  pFeature = 1,
  plot = "png",
  innerLinkage = "average",
  finalLinkage = "average",
  writeTable = FALSE,
  weightsItem = NULL,
  weightsFeature = NULL,
  verbose = FALSE,
  corUse = "everything"
)
}
\arguments{
\item{clusterNum}{A integer representing the return cluster number, this value should be less
than maxClusterNum(maxK). This is the only additional parameter in our function compared to the original
R package "ConsensusClusterPlus". All the other parameters are compatible to the function "ConsensusClusterPlus().}

\item{mRNAexp}{A matrix of mRNA expression}

\item{miRNAexp}{A matrix of miRNA expression}

\item{lncRNAexp}{A matrix of lncRNA expression}

\item{methylation}{A matrix of methylation}

\item{CNVmatrix}{A matrix of Copy number variation}

\item{maxK}{integer value. maximum cluster number  for Consensus Clustering Algorithm to evaluate.}

\item{clusterAlg}{character value. cluster algorithm. 'hc' heirarchical (hclust), 'pam' for paritioning around medoids, 'km' for k-means upon data matrix, 'kmdist' for k-means upon distance matrices (former km option), or a function that returns a clustering.}

\item{distance}{character value. 'pearson': (1 - Pearson correlation), 'spearman' (1 - Spearman correlation), 'euclidean', 'binary', 'maximum', 'canberra', 'minkowski" or custom distance function.}

\item{title}{character value for output directory. This title can be an absolute or relative path}

\item{reps}{integer value. number of subsamples(in other words, The iteration number of each cluster number)}

\item{pItem}{Please refer to the "ConsensusClusterPlus" package for detailed information.}

\item{pFeature}{Please refer to the "ConsensusClusterPlus" package for detailed information.}

\item{plot}{Please refer to the "ConsensusClusterPlus" package for detailed information.}

\item{innerLinkage}{Please refer to the "ConsensusClusterPlus" package for detailed information.}

\item{finalLinkage}{Please refer to the "ConsensusClusterPlus" package for detailed information.}

\item{writeTable}{Please refer to the "ConsensusClusterPlus" package for detailed information.}

\item{weightsItem}{Please refer to the "ConsensusClusterPlus" package for detailed information.}

\item{weightsFeature}{Please refer to the "ConsensusClusterPlus" package for detailed information.}

\item{verbose}{Please refer to the "ConsensusClusterPlus" package for detailed information.}

\item{corUse}{Please refer to the "ConsensusClusterPlus" package for detailed information.}
}
\value{
A list with the following elements.
\itemize{
 \item \strong{group} : A vector represent the group of cancer subtypes. The order is corresponding to the the samples in the data matrix.\cr

  This is the most important result for all clustering methods, so we place it as the first component. The format of group
  is consistent across different algorithms and therefore makes it convenient for downstream analyses. Moreover, the format
  of group is also compatible with the K-means result and the hclust (after using the cutree() function).

 \item \strong{distanceMatrix} : It is a sample similarity matrix. The more large value between samples in the matrix, the more similarity the samples are.

  We extracted this matrix from the algorithmic procedure because it is useful for similarity analysis among the samples based on the clustering results.

 \item \strong{originalResult} : The clustering result of the original function "ConsensusClusterPlus()"

 Different clustering algorithms have different output formats. Although we have the group component which has consistent format for all of the algorithms (making it easy for downstream analyses), we still keep the output from the original algorithms.
 }
}
\description{
This function is based on the R package "ConsensusClusterPlus".\cr
Based on the existing SNF algorithm, we improve it, increase the dataset type, and add the sample preprocessing function, because different sources of datasets, the sample format is not the same, researchers usually spend a lot of time in the unified sample format part, using our improved clustering algorithm,
users do not need to unify the sample format, directly input into the clustering algorithm, greatly saving the user's time.
}
\details{
If the data is a list containing the matched mutli-genomics  data matrices like the input as "ExecuteiCluster()" and "ExecuteSNF()",
  we use "z-score" to normalize features for each data matrix first. Then all the normalized data matrices from the data list are concatenated
  according to samples. The concatenated data matrix is the samples with a long features (all features in the data list).
  Our purpose is to make convenient comparing the different method with same dataset format. See examples.
}
\examples{
data(mRNAexp)
data(miRNAexp)
data(lncRNAexp)
data(methylation)
data(CNVmatrix)
result=ExecuteCC(clusterNum=3,mRNAexp,miRNAexp,lncRNAexp,methylation,CNVmatrix,maxK=5,clusterAlg="hc",distance="pearson")
result$group

}
\references{
Monti, S., Tamayo, P., Mesirov, J., Golub, T. (2003) Consensus Clustering: A Resampling-Based Method for Class Discovery and Visualization of Gene Expression Microarray Data. Machine Learning, 52, 91-118.
Xu,Taosheng \email{taosheng.x@gmail.com}, Thuc Le \email{Thuc.Le@unisa.edu.au}
}
\seealso{
\code{ConsensusClusterPlus}
}
