% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cluster_method.R
\name{ExecuteiCluster}
\alias{ExecuteiCluster}
\title{Execute iCluster based on five datasets.}
\usage{
ExecuteiCluster(
  mRNAexp,
  miRNAexp,
  lncRNAexp,
  methylation,
  CNVmatrix,
  k,
  lambda = NULL,
  scale = TRUE,
  scalar = FALSE,
  max.iter = 10
)
}
\arguments{
\item{mRNAexp}{A matrix of mRNA expression}

\item{miRNAexp}{A matrix of miRNA expression}

\item{lncRNAexp}{A matrix of lncRNA expression}

\item{methylation}{A matrix of methylation}

\item{CNVmatrix}{A matrix of Copy number variation}

\item{k}{Number of subtypes for the samples}

\item{lambda}{Penalty term for the coefficient matrix of the iCluster model}

\item{scale}{Logical value. If true, the genomic features in the matrix is centered.}

\item{scalar}{Logical value. If true, a degenerate version assuming scalar covariance matrix is used.}

\item{max.iter}{maximum iteration for the EM algorithm}
}
\value{
A list with the following elements.
\itemize{
  \item \strong{group} : A vector represent the group of cancer subtypes. The order is corresponding to the the samples in the data matrix.\cr

  This is the most important result for all clustering methods, so we place it as the first component. The format of group
  is consistent across different algorithms and therefore makes it convenient for downstream analyses. Moreover, the format
  of group is also compatible with the K-means result and the hclust (after using the cutree() function).

 \item \strong{originalResult} : The clustering result of the original function "iCluster2()".

 Different clustering algorithms have different output formats. Although we have the group component which has consistent format for all of the algorithms (making it easy for downstream analyses), we still keep the output from the original algorithms.
 }
}
\description{
Shen (2009) proposed a latent variable regression with a lasso constraint for joint modeling of multiple omics
data types to identify common latent variables that can be used to cluster patient samples into biologically and clinically relevant disease subtypes.\cr
This function is based on the R package "iCluster".\cr
Based on the existing SNF algorithm, we improve it, increase the dataset type, and add the sample preprocessing function, because different sources of datasets, the sample format is not the same, researchers usually spend a lot of time in the unified sample format part, using our improved clustering algorithm,
users do not need to unify the sample format, directly input into the clustering algorithm, greatly saving the user's time.
}
\details{
For iCluster algorithm, it cannot process high-dimensional data, otherwise it is very very time-consuming or reports a mistake.
 Based on test, it could smoothly run for the matrix with around 1500 features. Normally it need feature selection step first to reduce feature number.
}
\examples{
data(mRNAexp)
data(miRNAexp)
data(lncRNAexp)
data(methylation)
data(CNVmatrix)
result=ExecuteiCluster(mRNAexp,miRNAexp,lncRNAexp,methylation,CNVmatrix, k=3, lambda=list(0.44,0.33,0.28))
result$group

}
\references{
Ronglai Shen, Adam Olshen, Marc Ladanyi. (2009). Integrative clustering of multiple genomic data types using a joint latent variable model with application to breast and lung cancer subtype analysis. Bioinformatics 25, 2906-2912.\cr
Ronglai Shen, Qianxing Mo, Nikolaus Schultz, Venkatraman E. Seshan, Adam B. Olshen, Jason Huse, Marc Ladanyi, Chris Sander. (2012). Integrative Subtype Discovery in Glioblastoma Using iCluster. PLoS ONE 7, e35236
Xu,Taosheng \email{taosheng.x@gmail.com}, Thuc Le \email{Thuc.Le@unisa.edu.au}
}
\seealso{
\code{\link{iCluster2}}
}
