% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cluster_method.R
\name{ExecuteSNF}
\alias{ExecuteSNF}
\title{Execute SNF(Similarity Network Fusion ) based on five datasets.}
\usage{
ExecuteSNF(
  mRNAexp,
  miRNAexp,
  lncRNAexp,
  methylation,
  CNVmatrix,
  clusterNum,
  K = 20,
  alpha = 0.5,
  t = 20,
  plot = TRUE
)
}
\arguments{
\item{mRNAexp}{A matrix of mRNA expression}

\item{miRNAexp}{A matrix of miRNA expression}

\item{lncRNAexp}{A matrix of lncRNA expression}

\item{methylation}{A matrix of methylation}

\item{CNVmatrix}{A matrix of Copy number variation}

\item{clusterNum}{A integer representing the return cluster number}

\item{K}{Number of nearest neighbors}

\item{alpha}{Variance for local model}

\item{t}{Number of iterations for the diffusion process}

\item{plot}{Logical value. If true, draw the heatmap for the distance matrix with samples ordered to form clusters.}
}
\value{
A list with the following elements.
\itemize{
 \item \strong{group} : A vector represent the group of cancer subtypes. The order is corresponding to the the samples in the data matrix.\cr

  This is the most important result for all clustering methods, so we place it as the first component. The format of group
  is consistent across different algorithms and therefore makes it convenient for downstream analyses. Moreover, the format
  of group is also compatible with the K-means result and the hclust (after using the cutree() function).

 \item \strong{distanceMatrix} : It is a sample similarity matrix. The more large value between samples in the matrix, the more similarity the samples are.

  We extracted this matrix from the algorithmic procedure because it is useful for similarity analysis among the samples based on the clustering results.

 \item \strong{originalResult} : The clustering result of the original SNF algorithm"

 Different clustering algorithms have different output formats. Although we have the group component which has consistent format for all of the algorithms (making it easy for downstream analyses), we still keep the output from the original algorithms.
 }
}
\description{
Based on the existing SNF algorithm, we improve it, increase the dataset type, and add the sample preprocessing function, because different sources of datasets, the sample format is not the same, researchers usually spend a lot of time in the unified sample format part, using our improved clustering algorithm,
users do not need to unify the sample format, directly input into the clustering algorithm, greatly saving the user's time.
}
\examples{
data(mRNAexp)
data(miRNAexp)
data(lncRNAexp)
data(methylation)
data(CNVmatrix)
result=ExecuteSNF(mRNAexp,miRNAexp,lncRNAexp,methylation,CNVmatrix, clusterNum=3, K=20, alpha=0.5, t=20)
result$group

}
\references{
B Wang, A Mezlini, F Demir, M Fiume, T Zu, M Brudno, B Haibe-Kains, A Goldenberg (2014) Similarity Network Fusion: a fast and effective method to aggregate multiple data types on a genome wide scale. Nature Methods. Online. Jan 26, 2014
Xu,Taosheng \email{taosheng.x@gmail.com}, Thuc Le \email{Thuc.Le@unisa.edu.au}
}
