% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{lncRNAexp}
\alias{lncRNAexp}
\title{Dataset: lncRNA expression}
\format{
A data matrix
}
\description{
A glioblastoma (GBM) lncRNA expression dataset downloaded from TCGA.
}
\details{
\itemize{
 \item Rows are lncRNAs
 \item Columns are cancer samples
}
}
\examples{
data(lncRNAexp)
}
\keyword{datasets}
