% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sel.R
\name{metabolism}
\alias{metabolism}
\title{Select Metabolisms}
\usage{
metabolism(Data, type = "Lipid")
}
\arguments{
\item{Data}{A matrix representing the genomic data such as gene expression data.}

\item{type}{Type of metabolism of interest(Such as Lipid Metabolism,Carbohydrate Metabolism,Amino_acid Metabolism)}
}
\value{
A set of genes containing specific metabolic pathways
}
\description{
Select the type of metabolism you are interested in for downstream data analysis.
}
\examples{
data(mRNAexp)
data_lipid=metabolism(mRNAexp,type="Lipid")

}
