% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{methylation}
\alias{methylation}
\title{Dataset: methylation}
\format{
A data matrix
}
\description{
A methylation dataset downloaded from TCGA.
}
\details{
\itemize{
 \item Rows are features
 \item Columns are cancer samples
}
}
\examples{
data(methylation)
}
\keyword{datasets}
