% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{miRNAexp}
\alias{miRNAexp}
\title{Dataset: miRNA expression}
\format{
A data matrix
}
\description{
A glioblastoma (GBM) miRNA expression dataset downloaded from TCGA.
}
\details{
\itemize{
 \item Rows are miRNAs
 \item Columns are cancer samples
}
}
\examples{
data(miRNAexp)
}
\keyword{datasets}
