% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{mRNAexp}
\alias{mRNAexp}
\title{Dataset: mRNA expression}
\format{
A data matrix
}
\description{
A glioblastoma (GBM) protein coding gene expression dataset downloaded from TCGA.
}
\details{
\itemize{
 \item Rows are genes
 \item Columns are cancer samples
}
}
\examples{
data(mRNAexp)
}
\keyword{datasets}
