% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Valiation_and_visualization.R
\name{PACplot}
\alias{PACplot}
\title{Plot optimal K}
\usage{
PACplot(dot_df)
}
\arguments{
\item{dot_df}{A dataset containing K,DeltaArea,PAC columns.}
}
\value{
A figure.
}
\description{
Plot a figure for optimal K.
}
\examples{
data(dot_df)
PACplot(dot_df)

}
