% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{pathway_all}
\alias{pathway_all}
\title{Dataset: All metabolism geneset}
\format{
A data matrix
}
\description{
A metabolism dataset downloaded from the literature.
}
\examples{
data(pathway_all)
}
\references{
Peng X, Chen Z, Farshidfar F, Xu X, Lorenzi PL, Wang Y, Cheng F, Tan L, Mojumdar K, Du D, Ge Z, Li J, Thomas GV, Birsoy K, Liu L, Zhang H, Zhao Z, Marchand C, Weinstein JN; Cancer Genome Atlas Research Network, Bathe OF, Liang H. Molecular Characterization and Clinical Relevance of Metabolic Expression Subtypes in Human Cancers. Cell Rep. 2018 Apr 3;23(1):255-269.e4. doi: 10.1016/j.celrep.2018.03.077. PMID: 29617665; PMCID: PMC5916795.
}
\keyword{datasets}
