% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_preprocessing.R
\name{quantileNormalizeByFeature}
\alias{quantileNormalizeByFeature}
\title{Quantile Normalize By Feature}
\usage{
quantileNormalizeByFeature(matrix_to_normalize, target_distribution_matrix)
}
\arguments{
\item{matrix_to_normalize}{A matrix (m x n) with m samples as rows, and n features as columns.}

\item{target_distribution_matrix}{matrix (m2 x n) with m2 samples as rows, and n features as columns to use as the target distribution.}
}
\value{
A normalized matrix
}
\description{
A cross-platform normalization method using FSQN to improve comparability of DNA microarray and RNA-seq datasets
}
\examples{
set.seed(7)
target <- matrix(rnorm(100*150, mean = 1, sd = 1), nrow = 100, ncol = 150)
test <- matrix(rnorm(30*150, mean = 2, sd = 2), nrow = 30, ncol = 150)
normalized_test <- quantileNormalizeByFeature(test, target)

}
\references{
Franks JM, Cai G, Whitfield ML. Feature specific quantile normalization enables cross-platform classification of molecular subtypes using gene expression data. Bioinformatics. 2018 Jun 1;34(11):1868-1874. doi: 10.1093/bioinformatics/bty026. PMID: 29360996; PMCID: PMC5972664.
}
