% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_of_cluster.R
\name{Cluster_of_cluster}
\alias{Cluster_of_cluster}
\title{Cluster of cluster}
\usage{
Cluster_of_cluster(
  mRNAexp,
  miRNAexp,
  lncRNAexp,
  methylation,
  maxK = 6,
  reps = 50,
  pItem = 0.8,
  pFeature = 1,
  clusterAlg = "hc",
  distance = "pearson",
  innerLinkage = "complete",
  seed = 1262118388.71279,
  plot = "pdf"
)
}
\arguments{
\item{mRNAexp}{A matrix of mRNA expression}

\item{miRNAexp}{A matrix of miRNA expression}

\item{lncRNAexp}{A matrix of lncRNA expression}

\item{maxK}{integer value. maximum cluster number  for Consensus Clustering Algorithm to evaluate.}

\item{reps}{integer value. number of subsamples(in other words, The iteration number of each cluster number)}

\item{pItem}{Please refer to the "ConsensusClusterPlus" package for detailed information.}

\item{pFeature}{Please refer to the "ConsensusClusterPlus" package for detailed information.}

\item{clusterAlg}{character value. cluster algorithm. 'hc' heirarchical (hclust), 'pam' for paritioning around medoids, 'km' for k-means upon data matrix, 'kmdist' for k-means upon distance matrices (former km option), or a function that returns a clustering.}

\item{distance}{character value. 'pearson': (1 - Pearson correlation), 'spearman' (1 - Spearman correlation), 'euclidean', 'binary', 'maximum', 'canberra', 'minkowski" or custom distance function.}

\item{innerLinkage}{Please refer to the "ConsensusClusterPlus" package for detailed information.}

\item{seed}{optional numerical value. sets random seed for reproducible results.}

\item{plot}{Please refer to the "ConsensusClusterPlus" package for detailed information.}
}
\value{
A list of length maxK. Each element is a list containing consensusMatrix (numerical matrix), consensusTree (hclust), consensusClass (consensus class asssignments).
}
\description{
ClusterofCluster is a multi-method based on the R package "ConsensusClusterPlus".
}
\examples{
data(mRNAexp)
data(miRNAexp)
data(lncRNAexp)
data(methylation)
mRNAexp=data.filter(mRNAexp,percentage=0.6)
miRNAexp=data.filter(miRNAexp,percentage=0.6)
lncRNAexp=data.filter(lncRNAexp,percentage=0.6)
mRNAexp[mRNAexp==0]<-NA
miRNAexp[miRNAexp==0]<-NA
lncRNAexp[lncRNAexp==0]<-NA
mRNAexp<-as.matrix(mRNAexp)
miRNAexp<-as.matrix(miRNAexp)
lncRNAexp<-as.matrix(lncRNAexp)
mRNAexp=data.imputation(mRNAexp,fun="knn")
miRNAexp=data.imputation(miRNAexp,fun="knn")
lncRNAexp=data.imputation(lncRNAexp,fun="knn")
results = Cluster_of_cluster(mRNAexp,miRNAexp,lncRNAexp,methylation,maxK=6,reps=50,pItem=0.8,pFeature=1,clusterAlg="hc",distance="pearson",innerLinkage="complete",seed=1262118388.71279,plot="pdf")

}
\seealso{
\code{ConsensusClusterPlus}
}
